allow redis via ssl to specify a trust manager

+++ core/cas-server-core-tickets/src/test/java/org/apereo/cas/ticket/registry/BaseTicketRegistryTests.java
@@ -146,7 +146,9 @@ public abstract class BaseTicketRegistryTests {
         }
         ticketRegistry = this.getNewTicketRegistry();
         if (ticketRegistry != null) {
-            ticketRegistry.deleteAll();
+            if (!info.getTags().contains("SkipClearingTicketRegistry")) {
+                ticketRegistry.deleteAll();
+            }
             setUpEncryption();
         }
     }
